requirements.txt
torch>=2.1.0
transformers>=4.46.3
datasets>=2.16.0
peft>=0.7.0
trl>=0.7.10
bitsandbytes>=0.41.0
accelerate>=0.26.0
scipy
wandb==0.16.6
python-dotenv
pyyaml
